Add support for countries as choices in DRF serializer
Switching the CountryField field serializer to subclass ChoiceField
rather than a plain Field adds support for the choices argument,
which in turn means that the OPTIONS request will return all of
the supported countries. This is useful for JS applications where
the metadata API is used to pre-populate / validate input values.