Fix for CountryField multiple=true option and Django Admin View
An error occurred when a Countries field with the optional multiple=True field was used within DjangoAdmin.

The admin screen would fail to render the view, citing that the Country object was not iterable. It seems that there were problems with blank values being returned.

This commit adds checks to make sure that values which are set to None are handled properly.