Fix translation of common names
The translation of a string to a different language may very well be
the same string. For example, the German translation of the English
word "Venezuela" is "Venezuela".

This means, we cannot rely on string equality as a way to check whether
a translation for a string is available.

Therefore, let's remove the usage of fallback names as suggested by
@SmileyChris.

Fixes: https://github.com/SmileyChris/django-countries/issues/364