Fix TypeError in CountryFilter.choices()
When no country is selected, self.used_parameters.get(self.field.name)
will return None. This raises a TypeError when CountryFilter.choices()
tries to check whether a specific country choice is currently selected.

Avoid this by checking that the parameter is not empty similarly to how
Django's built-in admin.filters.RelatedFieldListFilter checks for this.

Fixes: 04662d1e282d ("Support Django 5.0")